refactor: factor out singleview
