+++ src/lib/singleview/productvulnerabilities/productvulnerabilities.ts
